Docstrings

SymPyPythonCall.SymPyPythonCall

SymPyPythonCall.sympy_plotting

SymPyPythonCall.Differential

SymPyPythonCall.VectorField

Base.match

Base.replace

SymPyPythonCall.Wild

SymPyPythonCall.asSymbolic

SymPyPythonCall.ask

SymPyPythonCall.lambdify

SymPyPythonCall.plot_implicit

SymPyPythonCall.plot_parametric_surface

SymPyPythonCall.refine

SymPyPythonCall.SymPyPythonCall — Module

SymPyPythonCall

Module to call Python's SymPy package from Julia using PythonCall

SymPyPythonCall.sympy_plotting — Constant

Plotting of symbolic objects.

The Plots package provide a uniform interface to many of Julia's plotting packages. SymPy plugs into Plots' "recipes."

The basic goal is that when Plots provides an interface for function objects, this package extends the interface to symbolic expressions.

In particular:

plot(ex::Sym, a, b; kwargs...) will plot a function evaluating ex over [a,b]

Example. Here we use the default backend for Plots to make a plot:

using Plots
@syms x
plot(x^2 - 2x, 0, 4)

plot(ex1, ex2, a, b; kwargs...) will plot the two expressions in a parametric plot over the interval [a,b].

Example:

@syms x
plot(sin(2x), cos(3x), 0, 4pi) ## also

For a few backends (those that support :path3d) a third symbolic expression may be added to have a 3d parametric plot rendered:

plot(sin(x), cos(x), x, 0, 4pi) # helix in 3d

plot(xs, ys, expression) will make a contour plot (for many backends).

@syms x y
plot(range(0,stop=5, length=50), range(0,stop=5, length=50), x*y)

To plot the surface z=ex(x,y) over a region we have Plots.surface. For example,

@syms x y
surface(-5:5, -5:5, 25 - x^2 - y^2)

a vectorfield plot can (inefficiently but directly) be produced following this example:

function vfieldplot(fx, fy; xlim=(-5,5), ylim=(-5,5), n=8)
    xs = range(xlim[1], stop=xlim[2], length=n)
    ys = range(ylim[1], stop=ylim[2], length=n)

    us = vec([x for x in xs, y in ys])
    vs = vec([y for x in xs, y in ys])
    fxs = vec([fx(x,y) for x in xs, y in ys])
    fys = vec([fy(x,y) for x in xs, y in ys])

    mxs = maximum(abs.(filter(!isinf, filter(!isnan, fxs))))
    mys = maximum(abs.(filter(!isinf, filter(!isnan, fys))))
    d = 1/2 * max((xlim[2]-xlim[1])/mxs, (ylim[2]-ylim[1])/mys) / n

    quiver(us, vs, quiver=(fxs.*d, fys.*d))

end
fx = (x + y) / sqrt(x^2 + y^2)
fy = (x - y) / sqrt(x^2 + y^2)
vfieldplot(fx, fy)

To plot two or more functions at once, the style plot([ex1, ex2], a, b) does not work. Rather, use plot(ex1, a, b); plot!(ex2), as in:

@syms x
plot(sin(x), 0, 2pi)
plot!(cos(x))

Some graphics provided by SymPy are available if PyPlot is installed, such as:

sympy.plotting.plot3d_parametric_surface

sympy.plotting.plot_implicit

Plot the parametrically defined surface [exs[1](u,v), exs[2](u,v), exs[3](u,v)] over [a0,a1] x [b0,b1]. The specification of the variables uses a tuple of the form (Sym, Real, Real) following the style of SymPy in integrate, say, where disambiguation of variable names is needed.

@syms theta, phi
r = 1
sympy.plotting.plot3d_parametric_surface((r*sin(theta)*sin(phi), r*sin(theta)*cos(phi), r*cos(theta)),
                        (theta, 0, pi), (phi, 0, pi/2))

sympy.plotting.plot_implicit(equation, (xvar, x0, x1), (yvar, y0, y1)) will plot implicitly the equation.

@syms x y
sympy.plotting.plot_implicit(Eq(x^2+ y^2,3), (x, -2, 2), (y, -2, 2))  # draw a circle

SymPyPythonCall.Differential — Type

Differential(x)

Use to find (partial) derivatives.

Example

@syms x y u()
Dx = Differential(x)
Dx(u(x,y))  # resolves to diff(u(x,y),x)
Dx(u)       # will evaluate diff(u(x), x)

SymPyPythonCall.VectorField — Type

VectorField(fx, fy): create an object that can be plotted as a vector field.

A vectorfield plot draws arrows at grid points proportional to [fx(x_i,y_i), fy(x_i,y_i)] to visualize the field generated by [fx, fy].

The plot command: plot(VectorField(fx, fy), xlims=(-5,5), ylims=(-5,5), n=8) will draw the vectorfield. This uses the default values, so the same graph would be rendered by plot(VectorField(fx,fy)).

To faciliate the visualization of solution to the ODE y' = F(x, y(x)), the call plot(VectorField(F)) will work. (The order is x then y, though often this is written as F(y(x),x).)

SymPy objects can be passed to VectorField, but this is a bit fragile, as they must each have two variables so that they can be called with two variables. (E.g., y(1,2) will be 1 not 2, as might be intended.)

Examples:

using Plots

fx(x,y) = sin(y); fy(x,y) = cos(y)
plot(VectorField(fx, fy), xlims=(-2pi, 2pi), ylims=(-2pi,2pi))

# plot field of y' = 3y*x over (-5,5) x (-5,5)
F(x,y) = 3*y*x
plot(VectorField(F))

# plot field and solution u' = u*(1-u)
@syms x u()
F(x,y) = y*(1-y)
out = dsolve(u'(x) - F(x, u(x)), x, (u, 0, 1))
plot(VectorField(F), xlims=(0,5), ylims=(0,2))
plot!(rhs(out))

Base.match — Method

match(pattern, expression, ...)

Match a pattern against an expression; returns a dictionary of matches.

If a match is unsuccesful, returns an empty dictionary. (SymPy returns "nothing")

The order of the arguments follows Julia's match function, not sympy.match, which can be used directly, otherwise.

Base.replace — Method

replace(expression, pattern, value, ...)

From: SymPy Docs

Traverses an expression tree and performs replacement of matching subexpressions from the bottom to the top of the tree. The default approach is to do the replacement in a simultaneous fashion so changes made are targeted only once. If this is not desired or causes problems, simultaneous can be set to false. In addition, if an expression containing more than one Wild symbol is being used to match subexpressions and the exact flag is true, then the match will only succeed if non-zero values are received for each Wild that appears in the match pattern.

Differences from SymPy:

"types" are specified via calling func on the head of an expression: func(sin(x)) -> sin, or directly through sympy.sin

functions are not supported

Examples (from the SymPy docs)

julia> using SymPyPythonCall

julia> @syms x, y, z
(x, y, z)

julia> f = log(sin(x)) + tan(sin(x^2)); print(f) # `print(f)` only so doctest can run
log(sin(x)) + tan(sin(x^2))

"type" -> "type"

Types are specified through func:

julia> func = SymPyPythonCall.Introspection.func
func (generic function with 1 method)

julia> replace(f, func(sin(x)), func(cos(x))) |> print  # type -> type
log(cos(x)) + tan(cos(x^2))

julia> #replace(f, sympy.sin, sympy.cos)  |>  print  # "log(cos(x)) + tan(cos(x^2))" **but fails**

julia> sin(x).replace(func(sin(x)), func(cos(x)), map=true)
(cos(x), Dict{Sym, Sym}(sin(x) => cos(x)))

"type" -> "function"

XXX

"pattern" -> "expression"

Using "Wild" variables allows a pattern to be replaced by an expression:

julia> a, b = Wild("a"), Wild("b")
(a_, b_)

julia> replace(f, sin(a), tan(2a)) |> print
log(tan(2*x)) + tan(tan(2*x^2))

julia> replace(f, sin(a), tan(a/2)) |> print
log(tan(x/2)) + tan(tan(x^2/2))

julia> f.replace(sin(a), a) |> print
log(x) + tan(x^2)

julia> (x*y).replace(a*x, a)
y

In the SymPy docs we have:

Matching is exact by default when more than one Wild symbol is used: matching fails unless the match gives non-zero values for all Wild symbols."

julia> replace(2x + y, a*x+b, b-a)  # y - 2
y - 2

julia> replace(2x + y, a*x+b, b-a, exact=false)  # y + 2/x
y + 2/x

"pattern" -> "func"

XXX

"func" -> "func"

XXX

SymPyPythonCall.Wild — Method

Wild(x)

create a "wild card" for pattern matching

SymPyPythonCall.asSymbolic — Method

asSymbolic(x)

Convert Python object into symbolic Julia object. Aliased to ↑.

SymPyPythonCall.ask — Method

ask(query)

Returns true, false, or nothing; ask

Example:

julia> using SymPyPythonCall

julia> @syms x::integer y::integer
(x, y)

julia> ask(𝑄.integer(x*y), sympy.And(𝑄.integer(x), 𝑄.integer(y)))
true

julia> filter(x -> ask(𝑄.prime(x)), 1:10) ## really slow isprime:
4-element Vector{Int64}:
 2
 3
 5
 7

SymPyPythonCall.lambdify — Function

lambdify(ex, vars=free_symbols();
         fns=Dict(), values=Dict, use_julia_code=false,
         invoke_latest=true)

Take a symbolic expression and return a Julia function or expression to build a function.

ex::Sym a symbolic expression with 0, 1, or more free symbols

vars a container of symbols to use for the function arguments. The default is free_symbols which has a specific ordering. Specifying vars allows this default ordering of arguments to be customized.

fns::Dict, vals::Dict: Dictionaries that allow customization of the function that walks the expression ex and creates the corresponding AST for a Julia expression. See SymPy.fn_map and SymPy.val_map for the default mappings of sympy functions and values into Julia's AST.

use_julia_code::Bool: use SymPy's conversion to an expression, the default is false

invoke_latest=true: if true will call eval and Base.invokelatest to return a function that should not have any world age issue. If false will return a Julia expression that can be evaled to produce a function.

Example:

julia> using SymPyPythonCall

julia> @syms x y z
(x, y, z)

julia> ex = x^2 * sin(x)
x^2*sin(x)

julia> fn = lambdify(ex);

julia> fn(pi)
0.0

julia> ex = x + 2y + 3z
x + 2*y + 3*z

julia> fn = lambdify(ex);

julia> fn(1,2,3) # order is by free_symbols
14

julia> ex(x=>1, y=>2, z=>3)
14

julia> fn = lambdify(ex, (y,x,z));

julia> fn(1,2,3)
13

!!! Note:

The default produces slower functions due to the calls to eval and Base.invokelatest. In the following g2 (which, as seen, requires additional work to compute) is as fast as calling f (on non symbolic types), whereas g1 is an order of magnitude slower in this example.

julia> @vars x
(x,)

julia> f(x) = exp(cot(x))
f (generic function with 1 method)

julia> g1 = lambdify(f(x))
#88 (generic function with 1 method)

julia> ex = lambdify(f(x), invoke_latest=false)
:(function var"##271"(x)
      exp(cot(x))
  end)

julia> @eval g2(x) = ($ex)(x)
g2 (generic function with 1 method)

An alternative, say, is to use GeneralizedGenerated's mk_function, as follows:

julia> using GeneralizedGenerated

julia> body = convert(Expr, f(x))
:(exp(cot(x)))

julia> g3 = mk_function((:x,), (), body)
function = (x;) -> begin
    (Main).exp((Main).cot(x))
end

This function will be about 2-3 times slower than f.

SymPyPythonCall.plot_implicit — Method

Plot an implicit equation

@syms x y
plot_implicit(Eq(x^2+ y^2,3), (x, -2, 2), (y, -2, 2))

SymPyPythonCall.plot_parametric_surface — Function

Render a parametrically defined surface plot.

Example:

@syms u, v
plot_parametric_surface((u*v,u-v,u+v), (u,0,1), (v,0,1))

This uses PyPlot, not Plots for now.

SymPyPythonCall.refine — Method

 refine

Simplify an expression using assumptions; refine.